fix(db): seed built-in roles, reject NUL bytes, cast COUNT, serialize job persists
- Seed built-in roles (admin, editor, user) at boot via ensureBuiltinRoles()
  with onConflictDoNothing, restoring data that legacy SQLite migration 0007
  provided via INSERT statements (the pg baseline is DDL-only).
- Reject NUL bytes in login credentials with 401 (postgres rejects \x00 in
  text columns; valid usernames never contain NUL, matching 1.x behavior).
- Cast COUNT(*)::int in user-files, audit-log, and roles listing queries so
  postgres returns a JS number instead of bigint-as-string.
- Serialize fire-and-forget job progress DB writes per jobId so the final
  "completed" status is never overwritten by a late-arriving "processing"
  write (race condition exposed by async postgres round-trips).

+++ apps/api/src/routes/roles.ts
@@ -60,7 +60,7 @@ export async function rolesRoutes(app: FastifyInstance): Promise<void> {
     const userCounts = await db
       .select({
         role: schema.users.role,
-        count: sql<number>`COUNT(*)`,
+        count: sql<number>`COUNT(*)::int`,
       })
       .from(schema.users)
       .groupBy(schema.users.role);
